refactor: Moved all components in dedicated folders
- The components have been moved into dedicated folders to improve
  the project structure.

- The components that represent the pages of the application
  were already in a 'Pages' folder. The folder has been
  renamed to 'pages' for consistency with all other folders
  in the project. The folder names are all lowercase.

- The components that are related to specific layouts have
  been moved into the 'layouts' folder.

- The components that represent the different context providers
  of the application have been moved into the 'providers' folder.

- The components that are used for routing have been moved into
  the 'routing' folder.

- All of these folders are inside of the generic 'components'
  folder. This makes it very simple to look for specific
  components based on their functionality in the application
  thanks to the dedicated folders.
