+++ frontend/src/components/layouts/AuthCard.jsx
