+++ frontend/src/components/layouts/VerticalDivider.jsx
